refactor: :bug: Measurement units
Changes in all the vh measurement units for adaptation of the page, in addition to correction of some measurements and function to go to different places on the page, the header nav does not work, however the footer nav does. I suspect the nav error is due to the filter in the header.